Rewrite in Zig (unfinished)
I do not get a chance to work on this often, when I do I actually
want to enjoy the experience. I no longer wish to deal with the
honestly horrific state of trying to write a professional audio
application across browsers.

While this is a rather drastic change, to an 'unproved' language
no less, its interfacing with C and considered language features
have made gaining access to an audio buffer much simpler.